backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.26.0
aiofiles>=23.2.1
python-frontmatter>=1.1.0
markdown>=3.5.0
cachetools>=5.3.0
docker>=7.0.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
